src/main.py
from langchain_openai import ChatOpenAI
from browser_use import Agent
from dotenv import load_dotenv
load_dotenv()

import asyncio

llm = ChatOpenAI(model="gpt-4o")

async def main():
    agent = Agent(
        task="Navigate to https://v0-basic-todo-app.vercel.app/, create a new todo item with the text 'Complete browser-use agent task', and print details about the HTML elements that were interacted with",
        llm=llm,
    )
    result = await agent.run()
    print(result)

if __name__ == "__main__":
    asyncio.run(main())
