pyproject.toml
[project]
name = "ai-assisted-acceptance-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.1.40",
    "langchain-openai>=0.3.1",
    "python-dotenv>=1.0.1",
]
